Abstracted most STFT functionality to the WaveformWrapper.
